@@ -1,8 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5.0</TargetFramework>
         <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
+        <LangVersion>default</LangVersion>
+        <Authors>Zoe Roux</Authors>
+        <Company>SDG</Company>
+        <PackageProjectUrl>https://github.com/AnonymusRaccoon/Kyoo.TheMovieDB</PackageProjectUrl>
     </PropertyGroup>
 
     <ItemGroup>
